+++ bravado_core/__init__.py
@@ -1,2 +1,2 @@
 # -*- coding: utf-8 -*-
-version = '4.13.4'
+version = '5.0.0'